+++ src/index.ts
@@ -6,6 +6,7 @@ export * from "./components/utilities/MapOrFallback.js";
 
 export * from "./components/Block.js";
 export * from "./components/Breadcrumbs.js";
+export * from "./components/Button.js";
 export * from "./components/ContentWrapper.js";
 export * from "./components/Header.js";
 export * from "./components/HumanDateTime.js";